requirements.txt
streamlit
google-generativeai
pandas
duckduckgo-search
PyPDF2
gTTS
matplotlib